refactor: Renomeia para diferenciar das entidades
